[Imp] Add SSE3 implementation for sample autotuner (for big samples, this doubles the processing speed for me) [Mod] Added command line switch /noAssembly which turns off MMX/SSE/3DNow!-accelerated code
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@3765 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/Autotune.h
@@ -23,7 +23,7 @@ protected:
 
 	SmpLength selectionStart, selectionEnd;
 
-	int8 *sampleData;
+	int16 *sampleData;
 	SmpLength sampleLength;
 
 public:
